refactor(FE): add pages folder
