+++ frontend/src/pages/UserProfile.tsx
@@ -1,7 +1,7 @@
+import { Button } from "@/components/ui/button";
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { useUser } from "@/hooks/useUser";
 import { useNavigate } from "react-router";
-import { Button } from "./ui/button";
-import { Card, CardContent, CardHeader, CardTitle } from "./ui/card";
 
 export default function UserProfile() {
   const { isPending, user } = useUser();
